refactor(test): move ALL  followup test files to test/topics/followup/ for  organization, updated imports
